+++ src/elements/document.rs
@@ -1,0 +1,5 @@
+use crate::elements::paragraph::Paragraph;
+#[derive(Debug, Default, Clone, PartialEq)]
+pub struct Document {
+    pub paragraphs: Vec<Paragraph>,
+}
