+++ src/elements/mod.rs
@@ -1,0 +1,9 @@
+mod document;
+mod paragraph;
+mod run;
+mod run_properties;
+
+pub use document::*;
+pub use paragraph::*;
+pub use run::*;
+pub use run_properties::*;
