+++ src/elements/paragraph.rs
@@ -1,0 +1,6 @@
+use crate::elements::run::Run;
+
+#[derive(Debug, Default, Clone, PartialEq)]
+pub struct Paragraph {
+    pub runs: Vec<Run>,
+}
